EDSC-4676: simplify mocking to only be in playwright
