+++ static/src/js/components/NlpSearchStatus/__tests__/NlpSearchStatus.test.tsx
@@ -132,7 +132,7 @@ describe('NlpSearchStatus component', () => {
           endDate: '2026-04-30T23:59:59.999Z'
         },
         spatial: {
-          polygon: ['-116.050002,44.358209,-116.050002,49.00139,-109.64514022973341,49.00139,-109.64514022973341,44.358209,-116.050002,44.358209']
+          boundingBox: ['-116.05,44.35821,44.35821,-109.64514,49.00139']
         }
       },
       selectedRegion: {}
